backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
whitenoise
gunicorn
psycopg2-binary
dj-database-url
cloudinary
django-cloudinary-storage
python-dotenv
Pillow